Move tests to new spec
